refactor: move env var manger to pkg/util
EnvVarManager is a wrapper interface for the
os operations to read and write environment variables.
This is useful for unit tests, giving the possibility
to mock the environment variables without affecting
the real os.
Currently, this is placed in the util package of the
virt-launcher.

Move the EnvVarManager to general pkg util, so that
it can be used by the whole pkg without importing
virt-launcher util.

Also, the `LookupEnv` method was added.

Signed-off-by: fossedihelm <ffossemo@redhat.com>