feat: Add machine-type-updater image
The machine type updater is a container that will iterate over
a vm list and will update their machine type.
Actually, the strategy to update a machine type is to remove it
from the vm spec, so that, once restarted, it will be valued
with the kubevirt default (the updated one).

The image reads environment variables:
- MACHINE_TYPE_GLOB: required, specify the pattern of the machine type that
 will be updated.
- NAMESPACE: required, filter the vm to be updated by namespace.
- RESTART_REQUIRED: optional, specify if the running vms should be
 immediately restarted.
- LABEL_SELECTOR: optional, filter the vm to be updated by a label
 selector.

If any of the required env variables are missing, or there is any
syntax error the container will exit with a specific error.

This image will be used from a job that will be spawned via virtctl
command `virtctl update machine-type <glob>`

Signed-off-by: fossedihelm <ffossemo@redhat.com>