refactoring: remove `utils.NewRandomVMWithDataVolumeWithRegistryImport`
`utils.NewRandomVMWithDataVolumeWithRegistryImport` is
used in the storage pkg, `clone_test` and `vm_test`.
In storage pkg it is widely used, while in the others it is just
some spot cases.
The removal from util package was addressed differently in
those cases.
In the storage pkg the function is preserved, making it unexported,
and changing the signature since:
- namespace variable was never used
- the accessMode was always ReadWriteOnce

In the clone_test and the vm_test the original function
content was directly expanded there.

Signed-off-by: fossedihelm <ffossemo@redhat.com>