vm_test: Refactor test_id:1521
Check for events rather than logs, which is well-defined API for users compared to
logs that is mostly available only to admins.
In order to check the message of the event and not only the reason,
the watcher pkg was enhanced with new functions that allow the users to
better specify the event they are looking for.
For future improvements, we can extend the watcher to look at the namespace
of the events.

Signed-off-by: fossedihelm <ffossemo@redhat.com>