bazel: Replace deprecated --features race with test specific race config
Removes the use of bazel test [..] --features race to avoid the
following warning:

```
WARNING: --features=race is no longer supported. Use
--io_bazel_rules_go//go/config:race instead.
```

`--io_bazel_rules_go//go/config:race` is not used instead as it
requires cgo to be used, this is only currently enabled for
pkg/virt-handler/node-labeller and so the race detector is enabled there
for now.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>