handler: ignore shutoff/unknown status/reason event
During startup process, the domain is in status Shutoff
with reason Unknown.

In normal cases the domain is so fast that this event
is not propagated to the handler.
In some cases, however, it could happen that the domain
is slower and this combination arrives to the handler.

In the latter case, the handler consider the domain not
alive, triggering the vmi deletion and final cleanup.

Under the hood, the vmi deletion is using the undefine of
the domain, which, if the domain is running,
it's converted to transient domain, without stopping it.

In such situation, hence, the domain is converted to transient,
but the final cleanup stop the listener, and unmount the volumes, etc.

The final outcome is that the launcher pod is in NotReady state,
since i.e. the volumecontainerdisk is gone.

Let's prevent the deletion, in case the Shutoff/Unknown combination
is triggered.

Signed-off-by: fossedihelm <ffossemo@redhat.com>